ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
pillow==10.2.0
numpy==1.26.3
python-multipart==0.0.6
